+++ tests/unit/test_set.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import unittest
 
 from gamspy import Alias
@@ -35,6 +37,12 @@ class SetSuite(unittest.TestCase):
         j1 = Set(self.m, "j")
         j2 = Set(self.m, "j")
         self.assertEqual(id(j1), id(j2))
+
+        # Set and domain containers are different
+        m = Container()
+        set1 = Set(self.m, "set1")
+        with self.assertRaises(GamspyException):
+            _ = Set(m, "set2", domain=[set1])
 
     def test_set_string(self):
         # Check if the name is reserved
@@ -83,9 +91,17 @@ class SetSuite(unittest.TestCase):
         )
 
     def test_records_assignment(self):
-        s = Set(self.m, "s")
+        new_cont = Container()
+        i = Set(self.m, "i")
+        j = Set(self.m, "j", domain=[i])
+        k = Set(new_cont, "k")
+
+        s = Set(self.m, "s", domain=[i])
         with self.assertRaises(TypeError):
             s.records = 5
+
+        with self.assertRaises(GamspyException):
+            j[k] = 5
 
     def test_set_operators(self):
         i = Set(self.m, "i", records=["seattle", "san-diego"])
@@ -181,7 +197,7 @@ class SetSuite(unittest.TestCase):
         t = Set(m, name="t", records=[f"t{i}" for i in range(1, 6)])
 
         sMinDown = Set(m, name="sMinDown", domain=[s, t])
-        sMinDown[s, t.lead((Ord(t) - Ord(s)))] = 1
+        sMinDown[s, t.lead(Ord(t) - Ord(s))] = 1
         self.assertEqual(
             m._unsaved_statements[-1].gamsRepr(),
             "sMinDown(s,t + (ord(t) - ord(s))) = 1;",